% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneMDBD.R
\name{get_loss_pointwise_tsne}
\alias{get_loss_pointwise_tsne}
\title{get_loss_pointwise_tsne}
\usage{
get_loss_pointwise_tsne(yy, Y, Ydist_sq, Mat)
}
\arguments{
\item{yy}{Embedding of the point that we want to calculate loss with.}

\item{Y}{Embedding matrix of other points.}

\item{Ydist_sq}{Distance matrix of embedding Y, can be calculated by as.matrix(dist(Y)**2).}

\item{Mat}{Simalarity matrix (P matrix) of all points (with the point we want to calculate loss with as the first point).}
}
\value{
The surrogate loss for the point yy.
}
\description{
get surrogate t-SNE loss for specified point
}
\examples{
get_loss_pointwise_tsne(yy, Y, Ydist_sq, Mat)
}
