% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneMDBD.R
\name{hessian_compute}
\alias{hessian_compute}
\title{hessian_compute}
\usage{
hessian_compute(Y, P)
}
\arguments{
\item{Y}{t-SNE embedding.}

\item{P}{Similarity matrix in t-SNE.}
}
\value{
Hessian matrix for t-SNE loss function at embedding Y.
}
\description{
computes the Hessian matrix for t-SNE loss function
}
\examples{
hessian_compute(Y, P)
}
