% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneMDBD.R
\name{plot_s_score_perplexity}
\alias{plot_s_score_perplexity}
\title{plot_s_score_perplexity}
\usage{
plot_s_score_perplexity(perplexity_candidates, s_score_mat, q = 0.95)
}
\arguments{
\item{perplexity_candidates}{Candidates for perplexities.}

\item{s_score_mat}{A matrix of singularity score for each data point at .}

\item{q}{Point size in the plot.}
}
\value{
A plot of t-SNE embedding with the quantile of singularity score as coloring.
}
\description{
plot singularity score
}
\examples{
plot_s_score_perplexity(perplexity_candidates, s_score_mat, q = 0.95)
}
