% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thePACKAGE.R
\name{plot_singularity_score}
\alias{plot_singularity_score}
\title{plot_singularity_score}
\usage{
plot_singularity_score(Y, s_score, point_size = 1)
}
\arguments{
\item{Y}{t-SNE embedding.}

\item{s_score}{Singularity score.}

\item{point_size}{Point size in the plot.}
}
\value{
A plot of t-SNE embedding with singularity score coloring.
}
\description{
plot singularity score
}
\examples{
plot_singularity_score(Y, s_score, point_size = 1)
}
